updates user model and adds cors to 'soundslikeobjectsrouter'

+++ routes/soundsLikeObjectsRouter.js
@@ -1,11 +1,13 @@
 const express = require('express');
 const SoundsLikeObject = require('../models/users/soundsLikeObject');
-
+const cors = require('./cors');
+const auth = require('../authenticate');
 const soundsLikeObjectRouter = express.Router();
 
 soundsLikeObjectRouter
 	.route('/')
-	.get((req, res, next) => {
+	.options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+	.get(cors.cors, auth.verifyUser, (req, res, next) => {
 		SoundsLikeObject.find()
 			.then((slos) => {
 				res.statusCode = 200;
@@ -14,7 +16,7 @@ soundsLikeObjectRouter
 			})
 			.catch((err) => next(err));
 	})
-	.delete((req, res, next) => {
+	.delete(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
 		SoundsLikeObject.deleteMany()
 			.then((respsone) => {
 				res.statusCode = 200;
@@ -23,9 +25,10 @@ soundsLikeObjectRouter
 			})
 			.catch((err) => next(err));
 	})
-	.post((req, res, next) => {
-		let slo = { ...req.body, userId: req.params.userId };
-		slo.soundsLike = slo.soundsLike.toLowerCase();
+	.post(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
+		let slo = { ...req.body, userId: req.user._id };
+		slo.title = slo.title.toLowerCase();
+
 		SoundsLikeObject.create(slo)
 			.then((SLO) => {
 				res.statusCode = 200;
@@ -36,28 +39,32 @@ soundsLikeObjectRouter
 	});
 
 // All SLOs not by user, but by artist users say they sound like. Artist name in endpoint should seperate multiple words in name with '-'.
-soundsLikeObjectRouter.route('/artist/:artist').get((req, res, next) => {
-	let artist = req.params.artist.split('-');
-	if (artist.length > 1) {
-		artist = artist.join(' ').toLowerCase();
-	} else {
-		artist = artist[0].toLowerCase();
-	}
-	console.log(artist);
+soundsLikeObjectRouter
+	.options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+	.route('/artist/:artist')
+	.get(cors.cors, auth.verifyUser, (req, res, next) => {
+		let artist = req.params.artist.split('-');
+		if (artist.length > 1) {
+			artist = artist.join(' ').toLowerCase();
+		} else {
+			artist = artist[0].toLowerCase();
+		}
+		console.log(artist);
 
-	SoundsLikeObject.find({ soundsLike: artist })
-		.then((results) => {
-			res.statusCode = 200;
-			res.setHeader('Content-Type', 'application/json');
-			res.json(results);
-		})
-		.catch((err) => next(err));
-});
+		SoundsLikeObject.find({ soundsLike: artist })
+			.then((results) => {
+				res.statusCode = 200;
+				res.setHeader('Content-Type', 'application/json');
+				res.json(results);
+			})
+			.catch((err) => next(err));
+	});
 
 // 'Sounds Like' Objects
 soundsLikeObjectRouter
 	.route('/user-collections/:userId')
-	.get((req, res, next) => {
+	.options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+	.get(cors.cors, auth.verifyUser, (req, res, next) => {
 		SoundsLikeObject.find({ userId: req.params.userId })
 			.then((SLOs) => {
 				res.statusCode = 200;
@@ -66,14 +73,62 @@ soundsLikeObjectRouter
 			})
 			.catch((err) => next(err));
 	})
-	.delete((req, res, next) => {
-		SoundsLikeObject.deleteMany({ userId: req.params.userId })
-			.then((response) => {
+	.delete(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
+		if (req.params.userId === req.user._id) {
+			SoundsLikeObject.deleteMany({ userId: req.params.userId })
+				.then((response) => {
+					res.statusCode = 200;
+					res.setHeader('Content-Type', 'application/json');
+					res.json(response);
+				})
+				.catch((err) => next(err));
+		} else {
+			res.statusCode = 403;
+			res.setHeader('Content-Type', 'application/json');
+			res.end('The files you are trying to delete do not belong to your account.');
+		}
+	})
+	.put((req, res) => {
+		res.statusCode = 405;
+		res.end('PUT not supported at this endpoint');
+	})
+	.post((req, res) => {
+		res.statusCode = 405;
+		res.end('POST not supported at this endpoint');
+	});
+
+soundsLikeObjectRouter
+	.route('/:soundsLikeID')
+	.options(cors.corsWithOptions, (req, res) => res.sendStatus(200))
+	.get(cors.cors, auth.verifyUser, (req, res, next) => {
+		SoundsLikeObject.findById(req.params.soundsLikeID)
+			.then((slo) => {
 				res.statusCode = 200;
 				res.setHeader('Content-Type', 'application/json');
-				res.json(response);
+				res.json(slo);
 			})
 			.catch((err) => next(err));
+	})
+	.delete(cors.corsWithOptions, auth.verifyUser, (req, res, next) => {
+		SoundsLikeObject.findById(req.params.soundsLikeID).then((slo) => {
+			console.log('SLO ID', slo);
+			console.log('USER ID', req.user);
+			if (req.user._id.equals(slo.userId)) {
+				SoundsLikeObject.findByIdAndDelete(req.params.soundsLikeID)
+					.then((respsone) => {
+						SoundsLikeObject.find({ userId: req.user._id }).then((list) => {
+							res.statusCode = 200;
+							res.setHeader('Content-Type', 'application/json');
+							res.json({ res: respsone, updatedList: list });
+						});
+					})
+					.catch((err) => next(err));
+			} else {
+				res.statusCode = 403;
+				res.setHeader('Content-Type', 'application/json');
+				res.end('You do not own the data you are trying to delete.');
+			}
+		});
 	})
 	.put((req, res) => {
 		res.statusCode = 405;
